backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
email-validator==2.1.0
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
Flask-Limiter==3.5.0
